embedding_indexing/requirements.txt
fastapi
httpx
pydantic
uvicorn
numpy
faiss-cpu
sentence-transformers==2.2.2
torch==1.13.1
